Simplify setup
This commit simplifies setup.py. Because `huepy/__init__.py` is a Python
file too, it can be imported in `setup.py`. This way we can use the
variables available in the init file. This makes determining the version
easier.  This commit does the same trick with `huepy/__init__.py` and
`huepy/hue.py`. It imports the hue source file, and generates a list for
all the created commands, no more hardcoding! :)